move log pattern analysis fetch code to common
